New test: github actions push/pull_request trigger
Baseline commit.
Things are not working how we'd like them to.

Goal: Always run tests once for pushed commit(s).
This should be the case for:
- push to master on this repo
- push to branch on this repo
- push to fork of this repo
- push to PR of this repo

Currently GHA is building twice on push to PR.
Once for PR and once for push.

I'd like to avoid this waste of compute.
Also, it is confusing.